autoconnect.py 0.3: save channels on exit, restructure
Previously it was not possible to differentiate between a '/disconnect
network' command and '/quit'. When you disconnect a network you might
not want to save the channel list, but you do want that on quit.

Also restructured the code a bit since the join and part callbacks
shared much code. Now also using a set to keep track of channels and
make add remove a bit easier.