tmux_env.py: support python 3.x and fixes
This commit replaces a `dict.iteritems()` call with the
py2/3-compatible `dict.items()` call. It also adds code to try the
`text=True` arg to the `subprocess.check_output` call so the output
can be handled as a `str` in python 3, falling back to the python
2-compatible call without the `text` arg if it catches and exception.
The updated args are stored in a global so we don't burn cycles
raising the exception every time `check_output` is called.

Other changes include minor formatting tweaks and replacing a
`+`-style string concatenation with inline string formatting.