colorize_nicks.py 28: fix colorization in input for colon completer
The global `VALID_NICK` regex allows the nick to contain a colon, however
the nick matching logic with stripping the first/last character is not
used within `#colorize_input_cb()`.
In combination with a colon ":" for `weechat.completion.nick_completer`,
this leads to nicks not being colorized in the input line.

To fix this, we can just always `rstrip()` the currently configured
`nick_completer` character.