fish.py 0.9.6: fix bug in key exchange again
Using s.encode() for hashing the DH key is not correct
as the resulting bytearray is utf-8 encoded and shouldn't.
The reuslting hash is then different from hashs produced by the
previous versions of this plugin, and probably other fish plugins (tested
with the original irssi one, and the key exchange didn't work either).
While using s.encode("latin-1") works, it's better to remove the encode() call
and rewrite int2bytes to actually return bytes (and not str). I used
python 3 built-ins for that, and modified bytes2int accordingly.
Fix bug in key exchange again

Using s.encode() for hashing the DH key is not correct
as the resulting bytearray is utf-8 encoded and shouldn't.
The reuslting hash is then different from hashs produced by the
previous versions of this plugin, and probably other fish plugins (tested
with the original irssi one, and the key exchange didn't work either).
While using s.encode("latin-1") works, it's better to remove the encode() call
and rewrite int2bytes to actually return bytes (and not str). I used
python 3 built-ins for that, and modified bytes2int accordingly. I also had
to modify dh1080_b64encode to make it return bytes, as it should be in py3.