Change urlgrab buffer to open the URL with <Return>
Urlgrab.py adds some keyboard shortcuts, to open a URL when the user is in the buffer. The current behaviour is that you have to use <alt><return>, changed to be just plain <return>. The shortcut is limited to this buffer AFAICT. Also added it to the documentation as it doesn't tell you there are keyboard shortcuts.