go.py 2.7: add new option to prefix short names with server names
The motivation here is that weechat-matrix uses machine-readable room
ids as full names that aren't human friendly. It does offer an option to
use human-readable full names, but these are neither unique nor stable,
causing trouble with logging. Therefore I figured that the best way to
fix this is to optionally use server names in go.py and let
weechat-matrix use the unique room ids.