add suppress_whitespace.py
`suppress_whitespace.py` can be used to remove the ability for
clients to send whitespace-only messages.

It can be used to avoid accidental messages in general.

If a user with an empty input buffer types in any whitespace
characters, it does not include them in the input buffer;
requiring some sort of valid input before whitespace is allowed.

Signed-off-by: Kevin Morris <kevr@0cost.org>