urlgrab.py 3.1: store urls by buffer `full_name`
This adds an option to store URLs by the full_name of the buffer, rather
than the short_name or name that is currently the default.

This fixes an issue that I had directly, where some buffers would
(frequently) get typing indicators at the short name, thus storing the
URL under `>general` instead of `#general`. This would break trying to
open URLs by index in the channel.

This should also address issue #90, as the full name should contain the
network information in it.

There is also a small linting cleanup, using `completion_list_add` and
fixing the weechat URLs