go.py: use buf number instead of name
This fixes some issues when switching to a buffer, when the buffer's
name is something that weechat does not like, e.g. "#Foo < Bar"

    weechat =!= Error with command "/buffer #Foo < Bar" (help on command: /help buffer)

These aren't really valid channel names(I think?), but some weechat
plugins (weechat-matrix) may create buffers with incompatible names like
that. In any case, switching based on the buffer *number* seems to be a
reliable compromise.