unwanted_msg.py 0.3: remove leading whitespace in every situation
Previously, this script was taking care of removing invalid spaces
used before a command. This patch improves the script to remove
leading whitespace characters in all cases, not just for a command.

Warnings have been removed with this commit. It merely does not allow
you to add leading whitespace into the input buffer, which feels
quite self-explanatory as a user (I think?).

In addition, the `/raw` command is now available. The `/raw` command
includes whitespace inside of enclosed delimiters which can be
configured (defaulted to []). See `/help raw` for details on
how it works and configuration.

Signed-off-by: Kevin Morris <kevr@0cost.org>