multiline.pl 0.6.4: support new key names in WeeChat 4.0.0
Uses the new key names for WeeChat >= 4.0.0 and the old key names for
older versions. For ctrl-m and ctrl-j it still uses these names rather
than the new return name, because it handles ctrl-m and ctrl-j
differently (but uses the correct version of it, e.g. ctrl-m for new
versions and ctrl-M for old versions, so it works in all versions).