buffer_autoclose.py 0.6: add setting to prefer closing some buffers
E.g. the Matrix script maps private rooms with multiple people to
buffers in weechat which are nominally not private. This means that
these rooms pile up, which is normally solved for real 1:1 private
chats.

Fix the problem by adding a way to autoclose also non-private buffers.

No behavior is changed unless the new config option is set.