go.py 3.0: refresh immediately buffer input when /go command is executed
With WeeChat >= 4.3.0, the signal "input_text_changed" is sent only when the
input text is actually changed.

When the /go command is bound to a key, we now have to manually force an update
of the bar item "input_text" so that the list of buffers is displayed
immediately, as it was with previous WeeChat versions.